change biginteger to long, update saveUser route

+++ src/main/java/com/ctw/ffn131330/base/BaseService.java
@@ -3,7 +3,6 @@ package com.ctw.ffn131330.base;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
-import java.math.BigInteger;
 import java.util.List;
 
 @Service
@@ -16,12 +15,12 @@ public abstract class BaseService <T extends BaseEntity> {
         return this.getRepository().save(entity);
     }
 
-    public T get(BigInteger id){
+    public T get(Long id){
         return this.getRepository().findById(id).orElse(null);
     }
 
     @Transactional
-    public void delete(BigInteger id){
+    public void delete(Long id){
         this.getRepository().deleteById(id);
     }
